requirements.txt
dash
pillow
numpy
flask